Tests controllres

+++ src/main/java/com/Digital/Digital/Infrastructure/Request/ProductsRequest/PriceRequest.java
@@ -12,6 +12,6 @@ public class PriceRequest {
     private Integer contractDuration;
     private Boolean oneTime;
     private String name;
-    private Double discount;
+    private Integer discount;
 
 }
